修改构建编译告警
# message auto-generated for no-merge-commit merge:
merge master_ad3 into master

清除构建编译告警

Created-by: insightDeveloper

Match-id-3655203f907dd1eae921df81eb7f5139ccbaaa77

+++ server/src/modules/base/core/JsonFileProcess.cpp
@@ -187,7 +187,7 @@ bool JsonFileProcess::SeekPhEndPosition(std::ifstream &file, bool endFlag, int b
     }
     int pos = -1;
     uint32_t leftCount = 0;
-    for (int i = offset; i < str.size(); ++i) {
+    for (size_t i = offset; i < str.size(); ++i) {
         if (str[i] == '{') {
             leftCount++;
             continue;
